+++ noodles-bgzf/src/async.rs
@@ -1,6 +1,7 @@
 //! Async BGZF.
 
 mod block_codec;
+pub mod fs;
 pub mod io;
 
 use self::block_codec::BlockCodec;